+++ pkg/analyzer/lib/src/dart/resolver/prefix_expression_resolver.dart
@@ -249,7 +249,6 @@ class PrefixExpressionResolver {
       }
       node.recordStaticType(staticType, resolver: _resolver);
     }
-    _resolver.nullShortingTermination(node);
   }
 
   void _resolveNegation(PrefixExpressionImpl node) {